Extend the status word to 64 bits
Well-behaved user_io should continue to work without any change.

+++ arc_file.h
@@ -12,6 +12,6 @@ void arc_reset();
 char *arc_get_rbfname();
 char *arc_get_corename();
 char *arc_get_conf();
-uint32_t arc_get_default();
+uint64_t arc_get_default();
 
 #endif // ARC_FILE_H
